Refactor AWS Batch models, DB, and add aio-redis implementation
